chore: 完成Upgrade30引擎基线迁移与服务端地址更新
1. 迁移至cocos2d-x-3.0-oh与CEGUI-0.7.9-r5引擎基线
2. 更新全量服务端配置地址为111.228.57.237
3. 修复Android平台编译适配与日志输出兼容
4. 调整HTTP客户端库与构建依赖路径
5. 添加CEGUI扩展组件与手势识别支持

+++ common/ljfm/code/source/ljfszipfile.cpp
@@ -33,7 +33,7 @@ namespace LJFMX
 		m_FullPathFileName = FI->GetFullPathFileName();
 		unsigned char* pData = NULL;
 #ifdef ANDROID
-		unsigned long ulSize = 0;
+		ssize_t ulSize = 0;
 		if (FI->m_FileArea == 0)
 		{
 			std::string PFN = ws2s(FI->GetFullPathFileName());
